requirements.txt
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.4.2
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pip==25.1
pycparser==2.21
pymongo==4.13.0
python-dotenv==1.1.1
requests==2.32.4
setuptools==78.1.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1